+++ Sources/SwifterSwift/UIKit/UICollectionViewExtensions.swift
@@ -147,6 +147,17 @@ public extension UICollectionView {
         register(UINib(nibName: identifier, bundle: bundle), forCellWithReuseIdentifier: identifier)
     }
 
+    /// SwifterSwift: Check whether IndexPath is valid within the CollectionView
+    ///
+    /// - Parameter indexPath: An IndexPath to check
+    /// - Returns: Boolean value for valid or invalid IndexPath
+    func isValidIndexPath(_ indexPath: IndexPath) -> Bool {
+        return indexPath.section >= 0 &&
+            indexPath.item >= 0 &&
+            indexPath.section < numberOfSections &&
+            indexPath.item < numberOfItems(inSection: indexPath.section)
+    }
+
 }
 
 #endif